// Frontend/src/components/UI/Minimal/types/user.ts
export type IAttachmentItem = {
  id?: number;
  entity_type: string;
  entity_id: number;
  file: string;
  file_original: string;
  file_size: number;
  created_date?: Date;
	deleted?: boolean;
};


// 배너
export type IBannerItem = {
  id?: number;
  entity_type: string;
  image_file: string;
  link_url: string;
  additional_text: string;
  created_date?: Date;
};

// 게시판
export type IBoardItem = {
  id?: number;
  admin_id?: number;
  member_id?: number;
  type: string;
  category?: string;
  subject: string;
  content: string;
  count: number;
  is_active: string;
  is_deleted: string;
  created_by_type: string;
  created_date: Date;
	writer_name?: string;
	writer_id?: string;
};
// ...
export type IMetaItem = {
  id?: number;
  entity_type: string;
  entity_id: string;
  entity_value: string;
  created_date?: Date | null;
	value?: string;
};